replace `from: {object}` with `from: courses` or `from: areas`
the object syntax was a holdover of when we would allow
from: {requirements: [list]}, which has been removed for quite a while
now. This is not only simpler to write, but I also took the opportunity
to clarify with the registrar that I don't need to handle repeated
courses beyond just removing them; they will flag repeats that should
not be used.