[MUSIC/BA] restrict "Performance Studies" to only L1/L2 courses
Clarified with the dept. that BAs aren't allowed to take L2 at all, but
sometimes will transition from BM→BA so we should accept them here.

The removal of 300-level is because of the Composition emphasis, which
requires MUSPF 326, and several terms of it, which entails a bunch of
extra combinations.